+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # QOwnNotes Changelog
 
+## 16.11.8
+- added information whether the notes path exists on the server to the
+  debug information
+
 ## 16.11.7
 - the tag color of the note will now be updated correctly if a tag is
   removed from the current note